Signpost

+++ src/main/java/br/com/grupobrasileiro/sistema_grupo_brasileiro_backend/dto/laio/user/view/EmployeeSimpleView.java
@@ -3,6 +3,6 @@ package br.com.grupobrasileiro.sistema_grupo_brasileiro_backend.dto.laio.user.vi
 public record EmployeeSimpleView(
         Long id,
         String fullName,
-        Integer avatar
+        Long avatar
 ) {
 }
